Group tools by domain
