+++ tests/agent-tool-boundaries.test.ts
@@ -21,26 +21,17 @@ describe("root and worker capability boundaries", () => {
     expect(toolFiles(rootTools)).toEqual([
       "agent.ts",
       "bash.ts",
-      "calendar/check-availability.ts",
-      "calendar/create-event.ts",
-      "calendar/list-events.ts",
+      "calendar.ts",
       "connection_search.ts",
-      "contacts/search.ts",
-      "gmail/read-thread.ts",
-      "gmail/search.ts",
-      "gmail/send.ts",
-      "gmail/update.ts",
+      "contacts.ts",
+      "gmail.ts",
       "load_skill.ts",
       "messaging.ts",
       "read_file.ts",
-      "request_vault_import.ts",
-      "request_vault_setup.ts",
-      "schedules/answer.ts",
-      "schedules/create.ts",
-      "schedules/list.ts",
-      "schedules/update.ts",
+      "schedules.ts",
       "todo.ts",
       "update_user_profile.ts",
+      "vault.ts",
       "write_file.ts",
     ]);
     expect(existsSync(`${rootTools}/sendMessage.ts`)).toBe(false);
