+++ tests/agent/capabilities.test.ts
@@ -1,40 +1,17 @@
 import type { DynamicResolveContext } from "eve/tools";
 import { describe, expect, it } from "vitest";
 import worker from "@/agent/subagents/worker/agent";
-import calendarCheckAvailabilityDefinition from "@/agent/tools/calendar/check-availability";
-import calendarCreateEventDefinition from "@/agent/tools/calendar/create-event";
-import calendarListEventsDefinition from "@/agent/tools/calendar/list-events";
-import contactsSearchDefinition from "@/agent/tools/contacts/search";
-import gmailReadThreadDefinition from "@/agent/tools/gmail/read-thread";
-import gmailSearchDefinition from "@/agent/tools/gmail/search";
-import gmailSendDefinition from "@/agent/tools/gmail/send";
-import gmailUpdateDefinition from "@/agent/tools/gmail/update";
+import calendar from "@/agent/tools/calendar";
+import contacts from "@/agent/tools/contacts";
+import gmail from "@/agent/tools/gmail";
 import messaging from "@/agent/tools/messaging";
-import requestVaultImport from "@/agent/tools/request_vault_import";
-import requestVaultSetup from "@/agent/tools/request_vault_setup";
-import answerSchedule from "@/agent/tools/schedules/answer";
-import createScheduleDefinition from "@/agent/tools/schedules/create";
-import listSchedulesDefinition from "@/agent/tools/schedules/list";
-import updateScheduleDefinition from "@/agent/tools/schedules/update";
+import schedules from "@/agent/tools/schedules";
 import updateUserProfile from "@/agent/tools/update_user_profile";
+import vault from "@/agent/tools/vault";
 
-const singletonTools = [
-  ["calendar-check-availability", calendarCheckAvailabilityDefinition],
-  ["calendar-create-event", calendarCreateEventDefinition],
-  ["calendar-list-events", calendarListEventsDefinition],
-  ["contacts-search", contactsSearchDefinition],
-  ["gmail-read-thread", gmailReadThreadDefinition],
-  ["gmail-search", gmailSearchDefinition],
-  ["gmail-send", gmailSendDefinition],
-  ["gmail-update", gmailUpdateDefinition],
-  ["request_vault_import", requestVaultImport],
-  ["request_vault_setup", requestVaultSetup],
-  ["schedules-answer", answerSchedule],
-  ["schedules-create", createScheduleDefinition],
-  ["schedules-list", listSchedulesDefinition],
-  ["schedules-update", updateScheduleDefinition],
-  ["update_user_profile", updateUserProfile],
-] as const;
+const singletonTools = [["update_user_profile", updateUserProfile]] as const;
+
+const groupedTools = [calendar, contacts, gmail, messaging, schedules, vault];
 
 describe("authored mode capability matrix", () => {
   it("gives interactive turns the authored coordinator capabilities", async () => {
@@ -92,13 +69,14 @@ async function authoredCapabilities(authenticator: string) {
   );
   capabilities.push(...resolvedSingletons.filter((name) => name !== null));
 
-  const resolveMessaging = messaging.events["turn.started"];
-  const resolvedMessaging = resolveMessaging
-    ? await resolveMessaging({}, context)
-    : null;
-  if (resolvedMessaging && !("execute" in resolvedMessaging)) {
-    capabilities.push(...Object.keys(resolvedMessaging));
-  }
+  const resolvedGroups = await Promise.all(
+    groupedTools.map(async (definition) => {
+      const resolve = definition.events["turn.started"];
+      const resolved = resolve ? await resolve({}, context) : null;
+      return resolved && !("execute" in resolved) ? Object.keys(resolved) : [];
+    })
+  );
+  capabilities.push(...resolvedGroups.flat());
 
   const resolveWorker = worker.events["turn.started"];
   if (resolveWorker && (await resolveWorker({}, context))) {
